// concert_project_next/src/types/Concert.ts
export interface Activity {
  id: number
  name: string
  latitude: number
  longitude: number
  description: string
  live: boolean
  broadcasted_live: string
  event: string
  mood: string
  type_name: string
  type_color: string
  type_color_description: string
  genre_name: string
  event_type_name: string
  price_category_name: string
  website: string
  address: string
  city: string
  phone: string
  email: string
  music: string
}